Add first version of modify_isolinux tool
This tool will enable us to easily change boot arguments of the boot.iso or
dvd.iso to avoid constant overriding of these. Just change this once to add
updates_image or use it repeatedly solve problem how to add overlay image.

Right now the tool know how to extract isolinux. That is not much helpful but
it's going to be extended.

We're re-using some functions from the create_dvd script which makes the
situation much simpler!