Use the latest tag if the Anaconda version cannot be detected
The `makeupdates` script is able to use the latest tag if it is run without
the `-t` option. However, the `update_image.py` doesn't support that and
tries to run the `show_version.sh` script to detect the version.

Let's change that and make the `show_version.sh` script optional:

* If the branch defines a specific Anaconda version, use it.
* If the script at `ShowVersionScriptPath` returns a version, use it.
* If there is no script or no version to use, use the latest tag.